backend/requirements.txt
fastapi[standard]>=0.115.12
uvicorn[standard]>=0.34.0
pydantic>=2.11.0
playwright>=1.49.0
beautifulsoup4>=4.12.0
colorthief>=0.2.1
pillow>=10.0.0
anthropic>=0.40.0
python-multipart>=0.0.20
aiofiles>=24.1.0
httpx>=0.28.0